\name{tpr-package}
\alias{tpr-package}
%\alias{tpr}
\docType{package}
\title{
Temporal Process Regression
}
\description{
  Fit regression models for temporal process responses
  with time-varying and time-independent coefficients.
}
\details{
\tabular{ll}{
Package: \tab tpr\cr
Type: \tab Package\cr
Version: \tab 0.0-1\cr
Date: \tab 2007-01-11\cr
License: \tab GPL 2.0 or later\cr
}
An overview of how to use the package, including the most important
functions
}
\author{
Jun Yan <jyan@stat.uiowa.edu>
}
\references{
  Fine, Yan, and Kosorok (2004): Temporal Process Regression. Biometrika.
}
\keyword{ package }
\seealso{
\code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
}
