\name{tpr-package}
\alias{tpr-package}
\docType{package}
\title{
Temporal Process Regression
}
\description{
  Fit regression models for temporal process responses
  with time-varying and time-independent coefficients.
}
\details{
An overview of how to use the package, including the most important
functions
}
\author{
Jun Yan <jun.yan@uconn.edu>
}
\references{
  Fine, Yan, and Kosorok (2004): Temporal Process Regression. Biometrika.
}
\keyword{ package }
